--- a/thunder/distributed/bucketing.py
+++ b/thunder/distributed/bucketing.py
@@ -140,7 +140,7 @@ class GradBuckets:
                 dist_prims.DistributedReduceOps.SUM,
                 group=group,
                 do_async=True,
-                skip_clone=False,
+                skip_clone=True,
             )
 
     def tell(self, grad: TensorProxy, group: ProcessGroup) -> None: